코드 분리 사용법
