+++ bucketlist_team15/lib/widget/deleteDialog.dart
@@ -1,0 +1,9 @@
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+Widget deleteDialog(){
+  return Dialog(
+
+  );
+}
